backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.7
chromadb==0.4.18
langchain-text-splitters==0.0.1
tiktoken==0.5.1
youtube-transcript-api==0.6.1
pydantic==2.5.0
python-dotenv==1.0.0
httpx>=0.25.0
yt-dlp>=2023.12.30 
